proper upload progress

+++ app/src/main/java/com/roplabs/bard/util/Helper.java
@@ -639,7 +639,7 @@ public class Helper {
                     @Override
                     public void onProgressChanged(int id, long bytesCurrent, long bytesTotal) {
                         if (bytesTotal > 0) {
-                            int percentage = (int) ((bytesCurrent/bytesTotal) * 100);
+                            int percentage = (int) ((bytesCurrent/(float) bytesTotal) * 100);
                             //Display percentage transfered to user
                             BardLogger.log("upload progress: " + bytesCurrent + " / " + bytesTotal);
                             progressDialog.setProgress(percentage);
